pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "bb-crypto-inventory"
version = "0.1.0"
description = "Blackbox Cryptographic Inventory Tool - LLM-agent-driven automated crypto asset discovery"
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
authors = [
    { name = "NyxFoundation" },
]
dependencies = [
    "click>=8.0",
    "httpx>=0.27",
    "openai>=1.0",
    "pydantic>=2.0",
    "cryptography>=43.0",
    "pyyaml>=6.0",
    "rich>=13.0",
    "python-nmap>=0.7",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "pytest-cov>=5.0",
    "ruff>=0.8",
    "mypy>=1.13",
]

[project.scripts]
bbci = "bbci.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["src/bbci"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.mypy]
python_version = "3.11"
strict = true

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
